Adds MiscAggregatedMetric and configured metrics (Recidiviz/recidiviz-data#19320)
## Description of the change

This change creates a new `MiscAggregatedMetric` subclass, which allows
users to store metadata about a metric that relies on an ad-hoc/manually
defined query rather than person_spans or person_events logic, and may
be calculated for only subsets of populations or aggregation levels
(e.g., `primary_district`, `avg_daily_caseload_officer`). Queries are
still defined in the `misc_aggregated_metrics` view builder, but this
change explicitly SELECTs only the metrics that are defined and
configured using the `MiscAggregatedMetric` class, making it so that
users must define a new metric object in order to add a misc. metric to
the pipeline.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19011

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 4887105aded3a05bd165167dbe770a609195ebc5