[Justice Counts] add warning if disaggregation column does not sum to total (Recidiviz/recidiviz-data#19289)
## Description of the change

In time_range_uploader.py, we now raise a warning if the user provided
bulk upload workbook contains breakdown sheets with values that do not
sum to the values in the aggregate sheets.

For example, if a user uploads a workbook for a prison system with total
admissions for 2021 as 100, but the admissions_by_type sheet sums to 95
for 2021, a warning will be thrown.

Example for annual metric:
<img width="727" alt="Screenshot 2023-03-16 at 4 25 05 PM"
src="https://user-images.githubusercontent.com/82831800/225756319-be7cf0cd-3634-4914-81a8-420daa09ea14.png">

Example for monthly metric:
<img width="729" alt="Screenshot 2023-03-16 at 4 25 18 PM"
src="https://user-images.githubusercontent.com/82831800/225756381-f572b682-4d88-470b-82bf-c3698d693a2d.png">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#15748

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: nichelle-hall <nichelle.a.hall@gmail.com>
GitOrigin-RevId: 478084814800392df04f1d8de079c2cfa100e1ca