Updating the BQ refresh completion node to a failure-sensitive operator (Recidiviz/recidiviz-data#19559)
## Description of the change

This replaces the BQ refresh completion node from a
`ShortCircuitOperator` that short-circuits and marks the DAG as success
if any of the BQ refreshes fail to mark the DAG as failed with an
`AirflowException`-throwing `PythonOperator` instead. This will force
automatic failure of the task and any downstream tasks tied to it.

<img width="1281" alt="Screenshot 2023-03-17 at 9 01 55 AM"
src="https://user-images.githubusercontent.com/5348488/225957147-141d50e8-e31c-4818-938f-ee3f144e9e7f.png">

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19537

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 179675f0d3e692abbe170e8e970e88f505674abc