[Bugfix] Fix issues with empty dataset deletion (Recidiviz/recidiviz-data#19578)
## Description of the change

Fixes crash where we were attempting to delete a dataset that had
routines and no tables which would throw a 400 because it wasn't empty
and could not be deleted.

Also, reduced the amount of time after a dataset is created that we are
willing to wait before we delete it if it is empty. Currently, if tables
in a dataset expire any time between 12pm PST and 12am PST, those
datasets will not be deleted when the deletion job runs at 12am PST.
Basically: any sandboxes created in the afternoon stick around for an
extra day, leaving around extra clutter.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19443

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9162a641916a7c57cb1f2baa95737ba0e52db40c