[US_MI] Fix offender_number joining for assessments (Recidiviz/recidiviz-data#19563)
## Description of the change

Starting sometime last month, it seems like the format of
`OffenderNumber` in the `ADH_SHOFFENDER` table changed such that there
are now leading zeroes. In order to make sure we can still correctly
join with the `ADH_OFFENDER` table (which has no leading zeroes), we
must trim all leading zeroes off of the `OffenderNumber` we're getting
from `ADH_SHOFFENDER`.

I noticed this issue when I saw that on the `Ingest View Results
Transformed to State Schema` table it said the most recently processed
rows for the assessments view was in February even though we've been
receiving MI tables in March. Since we use `ADH_OFFENDER` to get from
`offender number` to `offender id`, and then we use `offender id` to
inner join on `ADH_OFFENDER_BOOKING` we were generating no ingest view
results for any row that had an offender number with leading zeroes
(because it couldn't join onto anything in `ADH_OFFENDER_BOOKING`).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ae969e63fa486ac542033af1018ce43af143b9e5