Upgrade google-cloud-storage to >2.0.0 (Recidiviz/recidiviz-data#19601)
## Description of the change

I saw
[this](https://console.cloud.google.com/logs/query;cursorTimestamp=2023-03-17T22:24:04.517Z;query=resource.type%3D%22gae_app%22%20resource.labels.module_id%3D%22default%22%0A%0AlogName%3D%22projects%2Frecidiviz-staging%2Flogs%2Fapp%22%0Alabels.%22appengine.googleapis.com%2Ftrace_id%22%3D%22ee6ec51d6592cc4de8fead704aa8abcc%22%0Aseverity%3DERROR%0Atimestamp%3D%222023-03-17T22:24:04.513637Z%22%0AinsertId%3D%221f2fh21fkbg7b6%22;summaryFields=:false:32:beginning;timeRange=P4D?project=recidiviz-staging)
error in staging today where a request made by the GCS client was not
retried even though it got a (presumably retriable?) connection error.
It's possible that newer versions of the client properly retry on this
error, so I was inspired to try upgrading.

Tests / mypy appear to pass. Should we just deploy this and see what
happens? It's early in the week so easy to revert.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#4287

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: bf9c8851164eff9ab3666042bc709665d4cf77bd