[US_ND] Hydrate StateStaff (Recidiviz/recidiviz-data#19629)
## Description of the change
NOTE: This is a replacement for PR Recidiviz/recidiviz-data#19447, which was closed after the
requested changes were made.

Hydrates the StateStaff entity for supervision officers only in ND.
Hydrated from `docstars_officers` raw data. This is the first step in
addressing the issue of inactive officers appearing in Vitals in ND
(Recidiviz/recidiviz-data#19264).

While I was writing tests for this new view, I found that the fixture
files for `docstars_contacts_v2` ingest view tests contained PII. These
are handled in Recidiviz/recidiviz-data#19524

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6f5ee864cc32a30f2e7947273b1564ee5144372d