[US_ND] Clarify docstars_officers raw data descriptions (Recidiviz/recidiviz-data#19668)
## Description of the change

When I was looking for PO <> supervisor mapping information in ND raw
data, I noticed that the description for the `SUPERVISOR` field in the
`docstars_officers` raw data config file was inaccurate, and the
`SUPERVISOR_ID` field description could be more clear. These changes
clarify that `SUPERVISOR` is a Y/N field for whether or not the officer
_is_ a supervisor, and `SUPERVISOR_ID` is either the ID of their
supervisor, or the ID that represents them as a supervisor (if they are
one).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d8d84eb4dbb4398dd6b228f1ad8256bb13e88ea0