[Workflows] Unify EligibilityQueryConfig and OpportunityExportConfig  (Recidiviz/recidiviz-data#19666)
## Description of the change

Centralize information for an opportunity in one config. This logic
removes the need for an independent EligiblityQueryConfig that is used
in person records (i.e. `client_record` and `resident_record`) and the
`current_impact_funnel_status` view by adding the task completion type
to the `WorkflowsOpportunityConfig`. Also removes the need for the
`OpportunityExportConfig` by adding the source file and Firestore
collection name to the unified config.

I loaded impacted views to sandbox and can be found with `alexa_` prefix
. I also loaded comparison views to `alexa_comparison_` dataset

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-data/issues/19054

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 87f465deb4b28b99c461c58d7b5b52ee857dd3e7