[US_IX][Spotlight] Remap facilities in US_IX (Recidiviz/recidiviz-data#19720)
## Description of the change

This PR remaps facility names into groups to avoid the issue described
in Recidiviz/public-dashboardRecidiviz/recidiviz-data#583 also happening with Atlas data. This
has to be done separately from the US_IX mapping because the facility
names in the raw data are slightly different in Atlas.

This change also adds state-specific logic to the validation view to
exclude `NULL` facilities from the final result.

**Question for reviewers:** Can Idaho logic be dropped from these two
files? Are we still using or validating old Idaho data?

US_ID corollary fixes:
[Recidiviz/recidiviz-data#14376](https://github.com/Recidiviz/recidiviz-data/pull/14376): Remap
facilities in Idaho

[Recidiviz/recidiviz-data#14469](https://github.com/Recidiviz/recidiviz-data/pull/14469):Exclude
NULL facilities in Idaho

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18825

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2d3f025c9abd7befd8f380678429b9ea3850b63b